extract delegate into extension file
